### 제목 # 커밋 타입: 작업내용 (제목과 본문은 한 줄 띄워주세요) test: 내 주문내역 컨트롤러 테스트코드 작성 (my order preview) docs: my-order.adoc 생성 작성 refactor: like service/controller, grade controller 수정
### 본문 - 한 줄에 최대 72 글자까지만 입력하기
# 무엇을, 왜, 어떻게 했는지

# 꼬리말
# (선택) 이슈번호 작성

#   [커밋 타입]  리스트
#   feat      : 기능 (새로운 기능)
#   fix       : 버그 (버그 수정)
#   design    : CSS 등 사용자 UI 디자인 변경
#   refactor  : 리팩토링
#   style     : 스타일 (코드 형식, 세미콜론 추가: 비즈니스 로직에 변경 없음)
#   docs      : 문서 (문서 추가, 수정, 삭제)
#   test      : 테스트 (테스트 코드 추가, 수정, 삭제: 비즈니스 로직에 변경 없음)
#   chore     : 기타 변경사항 (빌드 스크립트 수정 등)
#   post      : 블로그 포스트 추가 (신규 포스트 작성 및 수정)
#   rename    : 파일 혹은 폴더명을 수정하거나 옮기는 작업만 하는 경우
#   remove    : 파일을 삭제하는 작업만 수행한 경우
# ------------------
#   [체크리스트]
#     제목 첫 글자는 대문자로 작성했나요?
#     제목은 명령문으로 작성했나요?
#     제목 끝에 마침표(.) 금지
#     제목과 본문을 한 줄 띄워 분리하기
#     본문에 여러줄의 메시지를 작성할 땐 "-"로 구분했나요?
# ------------------

+++ src/main/java/com/nhnacademy/_vidiabookstoreservice/user/service/impl/LikeServiceImpl.java
@@ -1,6 +1,8 @@
 package com.nhnacademy._vidiabookstoreservice.user.service.impl;
 
 import com.nhnacademy._vidiabookstoreservice.book.domain.Book;
+import com.nhnacademy._vidiabookstoreservice.book.exception.BookNotFoundException;
+import com.nhnacademy._vidiabookstoreservice.book.repository.BookRepository;
 import com.nhnacademy._vidiabookstoreservice.book.service.BookService;
 import com.nhnacademy._vidiabookstoreservice.user.domain.Like;
 import com.nhnacademy._vidiabookstoreservice.user.domain.User;
@@ -8,7 +10,9 @@ import com.nhnacademy._vidiabookstoreservice.user.dto.like.response.LikeResponse
 import com.nhnacademy._vidiabookstoreservice.user.dto.like.response.UserLikeResponse;
 import com.nhnacademy._vidiabookstoreservice.user.exception.AlreadyLikedException;
 import com.nhnacademy._vidiabookstoreservice.user.exception.LikeNotFoundException;
+import com.nhnacademy._vidiabookstoreservice.user.exception.UserNotFoundException;
 import com.nhnacademy._vidiabookstoreservice.user.repository.LikeRepository;
+import com.nhnacademy._vidiabookstoreservice.user.repository.UserRepository;
 import com.nhnacademy._vidiabookstoreservice.user.service.LikeService;
 import com.nhnacademy._vidiabookstoreservice.user.service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -25,14 +29,8 @@ public class LikeServiceImpl implements LikeService {
     private final UserService userService;
     private final BookService bookService;
     private final LikeRepository likeRepository;
-
-    /**
-     * 좋아요 여부확인
-     * */
-    @Override
-    public boolean isLiked(Long userId, Long bookId) {
-        return likeRepository.existsByUser_UserIdAndBook_Id(userId, bookId);
-    }
+    private final UserRepository userRepository;
+    private final BookRepository bookRepository;
 
     /**
      * 좋아요 리스트 조회
@@ -40,7 +38,9 @@ public class LikeServiceImpl implements LikeService {
     @Override
     @Transactional(readOnly = true)
     public List<LikeResponse> getLikes(Long userId) {
-        userService.getUserById(userId);
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException();
+        }
 
         List<Like> likeList = likeRepository.findAllByUser_UserId(userId);
         return likeList.stream()
@@ -56,7 +56,12 @@ public class LikeServiceImpl implements LikeService {
      * 좋아요 아이디 리스트 조회
      */
     @Override
+    @Transactional(readOnly = true)
     public List<UserLikeResponse> getLikeIdList(Long userId, List<Long> bookIds) {
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException();
+        }
+
         List<Like> likes = likeRepository.findAllByUser_UserIdAndBook_IdIn(userId, bookIds);
 
         return likes.stream()
@@ -70,7 +75,14 @@ public class LikeServiceImpl implements LikeService {
      * */
     @Override
     public void addLike(Long userId, Long bookId) {
-        if (isLiked(userId, bookId)) {
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException();
+        }
+        if (!bookRepository.existsById(bookId)) {
+            throw new BookNotFoundException(bookId);
+        }
+
+        if (likeRepository.existsByUser_UserIdAndBook_Id(userId, bookId)) {
             throw new AlreadyLikedException();
         }
 
@@ -90,6 +102,13 @@ public class LikeServiceImpl implements LikeService {
      * */
     @Override
     public void removeLike(Long userId, Long bookId) {
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException();
+        }
+        if (!bookRepository.existsById(bookId)) {
+            throw new BookNotFoundException(bookId);
+        }
+
         Like like = likeRepository.findByUser_UserIdAndBook_Id(userId, bookId)
                 .orElseThrow(LikeNotFoundException::new);
 
@@ -101,7 +120,11 @@ public class LikeServiceImpl implements LikeService {
      */
     @Override
     public void removeAllLike(Long userId, List<Long> bookIds) {
-        List<Like> likes = likeRepository.findAllByUser_UserId(userId);
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException();
+        }
+
+        List<Like> likes = likeRepository.findAllByUser_UserIdAndBook_IdIn(userId, bookIds);
         likeRepository.deleteAll(likes);
     }
 }
